+++ resources/views/components/google-button.blade.php
@@ -2,7 +2,7 @@
     'label' => 'Continue with Google',
 ])
 
-<a href="{{ route('oauth.redirect', 'google') }}" {{ $attributes->merge(['class' => 'gbtn']) }}>
+<a href="{{ route('oauth.redirect', 'google') }}" data-oauth-popup {{ $attributes->merge(['class' => 'gbtn']) }}>
     <span class="gbtn__icon" aria-hidden="true">
         <svg viewBox="0 0 48 48" width="18" height="18">
             <path fill="#FFC107" d="M43.611 20.083H42V20H24v8h11.303c-1.649 4.657-6.08 8-11.303 8-6.627 0-12-5.373-12-12s5.373-12 12-12c3.059 0 5.842 1.154 7.961 3.039l5.657-5.657C34.046 6.053 29.268 4 24 4 12.955 4 4 12.955 4 24s8.955 20 20 20 20-8.955 20-20c0-1.341-.138-2.65-.389-3.917z"/>